% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{add}
\alias{add}
\title{Return the sum of two numbers}
\usage{
add(a, b)
}
\arguments{
\item{a}{number one}

\item{b}{number two}
}
\value{
sum of two numbers
}
\description{
Return the sum of two numbers
}
\examples{
add(2,5)
}
